+++ sidm2/__init__.py
@@ -4,8 +4,8 @@ SID to SF2 Converter Package
 Converts Commodore 64 SID music files to SID Factory II format.
 """
 
-__version__ = "3.1.4"
-__build_date__ = "2026-03-21"
+__version__ = "3.1.6"
+__build_date__ = "2026-03-29"
 
 from .constants import *
 from .exceptions import *